Gemini bug fix #2
Make it so it correctly detects mod_config.xml changes outside of the application while its running and prompts the showConfirmModal when necessary. (Make sure to not overspam it of course, can only be one at a time so we stop checking while that pop-up is up. I am not saying that it does overspam it, but to double check that it does.) Refactor the code from the ground up when you begin seeing issues.